Move types to types file
