+++ src/commands/draw.ts
@@ -1,20 +1,6 @@
 import { Namespace } from "socket.io";
 import { CommandInterface } from "../commandController";
-
-export enum CoordinateType {
-  moveto,
-  lineto,
-}
-
-// TODO: Move shared types to shared types file
-export interface Coordinate {
-  x: number;
-  y: number;
-}
-
-export interface PathCoordinate extends Coordinate {
-  type: CoordinateType;
-}
+import { PathCoordinate, Coordinate, CoordinateType } from "../types";
 
 export class PathNode {
   coordinate: Coordinate;
@@ -84,12 +70,6 @@ class DrawPath {
     }
     return pathString;
   }
-}
-
-export interface ErasePath {
-  path: Coordinate[];
-  commandId: number;
-  threshold: number;
 }
 
 export class Drawing {
